+++ konduktor/dashboard/backend/main.py
@@ -1,25 +1,21 @@
-import datetime
-import time
 from typing import Any, Dict, List
 
-import requests
 from flask import Flask, jsonify, request
 from flask_cors import CORS
-from flask_socketio import SocketIO
 from kubernetes import client
 from kubernetes.client.exceptions import ApiException
 
 from konduktor.kube_client import batch_api, core_api, crd_api
+
+from .sockets import socketio
 
 app = Flask(__name__)
 
 # Ensure CORS is configured correctly
 cors = CORS(app, resources={r"/*": {"origins": "*"}})
 
-
-# SocketIO configuration
-socketio = SocketIO(app, cors_allowed_origins="*", ping_interval=25, ping_timeout=60)
-
+# Attach socketio to app after app is created
+socketio.init_app(app, cors_allowed_origins="*")
 
 # Use Kubernetes API clients
 # Initialize BatchV1 and CoreV1 API (native kubernetes)
@@ -27,15 +23,6 @@ batch_client = batch_api()
 core_client = core_api()
 # Initialize Kueue API
 crd_client = crd_api()
-
-# Global variables
-CLIENT_CONNECTED = False
-FIRST_RUN = True
-BACKGROUND_TASK_RUNNING = False
-LOG_CHECKPOINT_TIME = None
-SELECTED_NAMESPACES: list[str] = []
-PROD_LOGS_URL = "http://loki.loki.svc.cluster.local:3100/loki/api/v1/query_range"
-DEV_LOGS_URL = "http://localhost:3100/loki/api/v1/query_range"
 
 
 # Get a listing of workloads in kueue
@@ -87,31 +74,7 @@ def format_workloads(listing: Dict[str, Any]) -> List[Dict[str, Any]]:
     return res
 
 
-def format_log_entry(entry: List[str], namespace: str) -> Dict[str, str]:
-    """
-    Formats a log entry and its corresponding namespace
-
-    Args:
-        entry (List[str]): A list of log entry strings to be formatted.
-        namespace (str): The namespace to apply to each log entry.
-
-    Returns:
-        Dict[str, str]: an object with the following properties:
-        timestamp, log (message), and namespace
-    """
-    timestamp_ns = entry[0]
-    log_message = entry[1]
-    timestamp_s = int(timestamp_ns) / 1e9
-    dt = datetime.datetime.utcfromtimestamp(timestamp_s)
-    human_readable_time = dt.strftime("%Y-%m-%d %H:%M:%S")
-    formatted_log = {
-        "timestamp": human_readable_time,
-        "log": log_message,
-        "namespace": namespace,
-    }
-    return formatted_log
-
-
+"""
 # for testing: prints workloads in kueue
 def list_all_workloads(namespace="default"):
     try:
@@ -143,64 +106,7 @@ def list_all_jobs():
 
     except ApiException as e:
         print(f"Failed to list jobs: {e}")
-
-
-def get_logs(FIRST_RUN: bool, dev: bool) -> List[Dict[str, str]]:
-    global LOG_CHECKPOINT_TIME
-
-    # print(f'SELECTED NAMESPACES (GET_LOGS): {SELECTED_NAMESPACES}')
-
-    # Use the selected namespaces in the query
-    namespace_filter = (
-        "|".join(SELECTED_NAMESPACES) if SELECTED_NAMESPACES else "default"
-    )
-    query = f'{{k8s_namespace_name=~"{namespace_filter}"}}'
-
-    # print(f'QUERY (GET_LOGS): {query}')
-
-    if FIRST_RUN:
-        # Calculate how many nanoseconds to look back when first time looking at logs
-        # (currently 1 hour)
-        now = int(time.time() * 1e9)
-        one_hour_ago = now - int(3600 * 1e9)
-        start_time = str(one_hour_ago)
-    else:
-        # calculate new start_time based on newest, last message
-        if LOG_CHECKPOINT_TIME is None:
-            LOG_CHECKPOINT_TIME = 0
-        start_time = str(int(LOG_CHECKPOINT_TIME) + 1)
-
-    params = {"query": query, "start": start_time, "limit": "300"}
-
-    url = DEV_LOGS_URL if dev else PROD_LOGS_URL
-    response = requests.get(url, params=params)
-    formatted_logs = []
-
-    last = 0
-
-    if response.status_code == 200:
-        data = response.json()
-        rows = data["data"]["result"]
-
-        for row in rows:
-            namespace = row["stream"]["k8s_namespace_name"]
-            for value in row["values"]:
-                last = max(int(value[0]), last)
-                formatted_logs.append(format_log_entry(value, namespace))
-        # print('status 200 getting logs')
-
-    if formatted_logs:
-        # sort because sometimes loki API is wrong and logs are out of order
-        formatted_logs.sort(
-            key=lambda log: datetime.datetime.strptime(
-                log["timestamp"], "%Y-%m-%d %H:%M:%S"
-            )
-        )
-        LOG_CHECKPOINT_TIME = last
-
-    # print(f'formatted logs length: {len(formatted_logs)}')
-
-    return formatted_logs
+"""
 
 
 # ROUTES
@@ -316,56 +222,6 @@ def update_priority():
     except ApiException as e:
         print(f"Exception: {e}")
         return jsonify({"error": str(e)}), e.status
-
-
-# websocket connection for continuous log fetching
-
-
-@socketio.on("connect")
-def handle_connect():
-    global CLIENT_CONNECTED, FIRST_RUN, BACKGROUND_TASK_RUNNING
-    CLIENT_CONNECTED = True
-    FIRST_RUN = True
-    print("Client connected")
-
-    # Start the background task only if it's not already running
-    if not BACKGROUND_TASK_RUNNING:
-        BACKGROUND_TASK_RUNNING = True
-        socketio.start_background_task(send_logs)
-
-
-def send_logs():
-    global CLIENT_CONNECTED, FIRST_RUN, BACKGROUND_TASK_RUNNING
-    while CLIENT_CONNECTED:
-        logs = []
-        try:
-            # Attempt to get logs from the production setup; if fails, switch to dev
-            logs = get_logs(FIRST_RUN, False)
-        except Exception:
-            logs = get_logs(FIRST_RUN, True)
-
-        FIRST_RUN = False  # After the first successful fetch, set to False
-        if logs:
-            socketio.emit("log_data", logs)
-
-        time.sleep(5)
-
-    # Background task is no longer running after the loop
-    BACKGROUND_TASK_RUNNING = False
-
-
-@socketio.on("update_namespaces")
-def handle_update_namespaces(namespaces):
-    global SELECTED_NAMESPACES
-    SELECTED_NAMESPACES = namespaces
-
-
-@socketio.on("disconnect")
-def handle_disconnect():
-    global CLIENT_CONNECTED, FIRST_RUN
-    CLIENT_CONNECTED = False
-    FIRST_RUN = True
-    print("Client disconnected")
 
 
 if __name__ == "__main__":
